+++ master/docs/release-notes.rst
@@ -90,6 +90,9 @@ Features
 * ``Git`` has a new ``getDescription`` option, which will run `git describe` after checkout
   normally.  See the documentation for details.
 
+* A new ternary substitution operator ``:?:`` and ``:#?:`` to use with the ``Interpolate``
+  and ``WithProperties`` classes.
+
 Slave
 -----